Revert "hotbar ui FAIL"
This reverts commit 7619632654c565e785b4f9ff70fea19f15c48440.
