--- Common/UI/Hotbar/Components/HotbarSlot.cs
@@ -1,9 +1,0 @@
-﻿namespace PathOfTerraria.Common.UI.Hotbar.Components;
-
-internal abstract class HotbarSlot
-{
-	public virtual float GetFullWidth()
-	{
-		return 0f; // TODO
-	}
-}
