--- Common/UI/Hotbar/Components/Sets/BuildingSet.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace PathOfTerraria.Common.UI.Hotbar.Components.Sets;
-
-internal sealed class BuildingSet : HotbarSet
-{
-	private sealed class BuildingSpecialSlot : SpecialSlot
-	{
-	}
-
-	public override IEnumerable<HotbarSlot> HotbarSlots => throw new NotImplementedException();
-}
