--- Common/UI/Hotbar/Components/SpecialSlot.cs
@@ -1,6 +1,0 @@
-﻿namespace PathOfTerraria.Common.UI.Hotbar.Components;
-
-internal abstract class SpecialSlot : HotbarSlot
-{
-	public bool DrawingSelected { get; set; }
-}
